+++ packages/effector/index.d.ts
@@ -1,6 +1,5 @@
-//@flow
 
-import type {Stream} from 'most'
+import {Stream} from 'most'
 
 export type Subscriber<A> = {
  next(value: A): void,
@@ -13,18 +12,19 @@ export type Subscription = {
  unsubscribe(): void,
 }
 
-declare export class Event<E> {
+export class Event<E> {
  (payload: E): E;
- watch(watcher: (payload: E) => any): void;
+ watch(watcher: (payload: E) => any): Subscription;
  map<T>(fn: (_: E) => T): Event<T>;
  prepend<Before>(fn: (_: Before) => E): Event<Before>;
  subscribe(subscriber: Subscriber<E>): Subscription;
- to(store: Store<E>, _: void): void;
- to<T>(store: Store<T>, reducer: (state: T, payload: E) => T): void;
+ to(store: Store<E>, _: void): Subscription;
+ to<T>(store: Store<T>, reducer: (state: T, payload: E) => T): Subscription;
+ getType(): string;
  epic<T>(fn: (_: Stream<E>) => Stream<T>): Event<T>;
 }
 
-declare export class Effect<Params, Done, Fail = Error> {
+export class Effect<Params, Done, Fail = Error> {
  (
   payload: Params,
  ): {
@@ -32,19 +32,23 @@ declare export class Effect<Params, Done, Fail = Error> {
   fail(): Promise<{params: Params, error: Fail}>,
   promise(): Promise<Done>,
  };
- /*::+*/ done: Event<{params: Params, result: Done}>;
- /*::+*/ fail: Event<{params: Params, error: Fail}>;
- use(asyncFunction: (params: Params) => Promise<Done>): void;
- watch(watcher: (payload: Params) => any): void;
- prepend<Before>(fn: (_: Before) => Params): Event<Before>;
+ done: Event<{params: Params, result: Done}>;
+ fail: Event<{params: Params, error: Fail}>;
+ use: {
+  (asyncFunction: (params: Params) => Promise<Done>): void;
+  getCurrent(): (params: Params) => Promise<Done>;
+ };
+ watch(watcher: (payload: Params) => any): Subscription;
  //map<T>(fn: (_: E) => T): Event<T>,
+ prepend<Before>(fn: (_: Before) => Params): Event<Before>;
  subscribe(subscriber: Subscriber<Params>): Subscription;
- to(store: Store<Params>, _: void): void;
- to<T>(store: Store<T>, reducer: (state: T, payload: Params) => T): void;
+ to(store: Store<Params>, _: void): Subscription;
+ to<T>(store: Store<T>, reducer: (state: T, payload: Params) => T): Subscription;
  epic<T>(fn: (_: Stream<Params>) => Stream<T>): Event<T>;
+ getType(): string;
 }
 
-declare export class Store<State> {
+export class Store<State> {
  reset(event: Event<any> | Effect<any, any, any>): this;
  dispatch(action: any): any;
  getState(): State;
@@ -54,17 +58,18 @@ declare export class Store<State> {
  map<T>(fn: (_: State) => T): Store<T>;
  on<E>(
   event: Event<E> | Effect<E, any, any>,
-  handler: (state: $NonMaybeType<State>, payload: E) => ?State,
+  handler: (state: State, payload: E) => (State | void),
  ): this;
+ replaceReducer(_: any): any;
  subscribe(listner: any): Subscription;
- /*::+*/ watch: (<E>(
+ watch<E>(
   watcher: (state: State, payload: E, type: string) => any,
   _: void,
- ) => void) &
-  (<E>(
-   event: Event<E> | Effect<E, any, any>,
-   watcher: (state: State, payload: E, type: string) => any,
-  ) => void);
+ ): Subscription;
+ watch<E>(
+  event: Event<E> | Effect<E, any, any>,
+  watcher: (state: State, payload: E, type: string) => any,
+ ): Subscription;
  thru<U>(fn: (store: Store<State>) => U): U;
  epic<T, S>(
   event: Event<T> | Effect<T, any, any>,
@@ -73,59 +78,57 @@ declare export class Store<State> {
  displayName?: string;
 }
 
-declare export class Domain {
+export class Domain {
  event<Payload>(name: string): Event<Payload>;
  effect<Params, Done, Fail>(name: string): Effect<Params, Done, Fail>;
  domain(name: string): Domain;
  store<State>(defaultState: State): Store<State>;
 }
 
-declare export function createEvent<E>(eventName: string): Event<E>
+export function createEvent<E>(eventName: string): Event<E>
 
-declare export function createEffect<Params, Done, Fail>(
+export function createEffect<Params, Done, Fail>(
  effectName: string,
 ): Effect<Params, Done, Fail>
 
-declare export function createStore<State>(defaultState: State): Store<State>
-
-type MapStoreFields = (<O: {[string]: any}>(
- _: O,
-) => $ObjMap<O, <S>(field: Store<S> | S) => S>) &
- (<O>(_: O) => O)
+export function createStore<State>(defaultState: State): Store<State>
 
 declare export function createStoreObject<State>(
  defaultState: State,
-): Store<$Call<MapStoreFields, State>>
+): Store<any>
 
-declare export function createReduxStore<T>(
+export function createReduxStore<T>(
  reducer: (state: T, event: any) => T,
  preloadedState?: T,
  enhancer?: Function | Function[],
 ): Store<T>
 
-declare export function createDomain(domainName?: string): Domain
+export function createDomain(domainName?: string): Domain
 
-declare export function combine<R>(fn: () => R): Store<R>
-declare export function combine<A, R>(a: Store<A>, fn: (a: A) => R): Store<R>
-declare export function combine<A, B, R>(
+export function combine<R>(fn: () => R): Store<R>
+export function combine<A, R>(
+ a: Store<A>,
+ fn: (a: A) => R,
+): Store<R>
+export function combine<A, B, R>(
  a: Store<A>,
  b: Store<B>,
  fn: (a: A, b: B) => R,
 ): Store<R>
-declare export function combine<A, B, C, R>(
+export function combine<A, B, C, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
  fn: (a: A, b: B, c: C) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, R>(
+export function combine<A, B, C, D, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
  d: Store<D>,
  fn: (a: A, b: B, c: C, d: D) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, E, R>(
+export function combine<A, B, C, D, E, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
@@ -133,7 +136,7 @@ declare export function combine<A, B, C, D, E, R>(
  e: Store<E>,
  fn: (a: A, b: B, c: C, d: D, e: E) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, E, F, R>(
+export function combine<A, B, C, D, E, F, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
@@ -142,7 +145,7 @@ declare export function combine<A, B, C, D, E, F, R>(
  f: Store<F>,
  fn: (a: A, b: B, c: C, d: D, e: E, f: F) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, E, F, G, R>(
+export function combine<A, B, C, D, E, F, G, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
@@ -152,7 +155,7 @@ declare export function combine<A, B, C, D, E, F, G, R>(
  g: Store<G>,
  fn: (a: A, b: B, c: C, d: D, e: E, f: F, g: G) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, E, F, G, H, R>(
+export function combine<A, B, C, D, E, F, G, H, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
@@ -163,7 +166,7 @@ declare export function combine<A, B, C, D, E, F, G, H, R>(
  h: Store<H>,
  fn: (a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, E, F, G, H, I, R>(
+export function combine<A, B, C, D, E, F, G, H, I, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
@@ -175,7 +178,7 @@ declare export function combine<A, B, C, D, E, F, G, H, I, R>(
  i: Store<I>,
  fn: (a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, E, F, G, H, I, J, R>(
+export function combine<A, B, C, D, E, F, G, H, I, J, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
@@ -188,7 +191,7 @@ declare export function combine<A, B, C, D, E, F, G, H, I, J, R>(
  j: Store<J>,
  fn: (a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J) => R,
 ): Store<R>
-declare export function combine<A, B, C, D, E, F, G, H, I, J, K, R>(
+export function combine<A, B, C, D, E, F, G, H, I, J, K, R>(
  a: Store<A>,
  b: Store<B>,
  c: Store<C>,
@@ -202,3 +205,4 @@ declare export function combine<A, B, C, D, E, F, G, H, I, J, K, R>(
  k: Store<K>,
  fn: (a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, i: I, j: J, k: K) => R,
 ): Store<R>
+