+++ src/effector/index.h.js
@@ -27,6 +27,8 @@ export type Subscription = {
  unsubscribe(): void,
 }
 
+export type Unsubscribe = () => void
+
 export type GraphiteMeta = {
  next: MultiStep,
  seq: SeqStep,
@@ -36,12 +38,15 @@ export type Event<E> = {
  (payload: E): E,
  getType(): string,
  create(payload: E, type: string): E,
- watch(watcher: (payload: E) => any): () => void,
+ watch(watcher: (payload: E) => any): Subscription,
  map<T>(fn: (_: E) => T): Event<T>,
  prepend<Before>(fn: (_: Before) => E): Event<Before>,
  subscribe(subscriber: Subscriber<E>): Subscription,
- to(store: Store<E>, _: void): void,
- to<T>(store: Store<T>, reducer: (state: T, payload: E) => T): void,
+ //prettier-ignore
+ +to: (
+  & (<T>(store: Store<T>, reducer: (state: T, payload: E) => T) => Subscription)
+  & ((store: Store<E>, _: void) => Subscription)
+ ),
  epic<T>(fn: (_: Stream<E>) => Stream<T>): Event<T>,
  getType(): string,
  shortName: string,
@@ -59,13 +64,22 @@ export type Effect<Params, Done, Fail = Error> = {
  },
  done: Event<{params: Params, result: Done}>,
  fail: Event<{params: Params, error: Fail}>,
- use: (asyncFunction: (params: Params) => Promise<Done>) => void,
- watch(watcher: (payload: Params) => any): void,
+ use: {
+  (asyncFunction: (params: Params) => Promise<Done>): void,
+  getCurrent(): (params: Params) => Promise<Done>,
+ },
+ watch(watcher: (payload: Params) => any): Subscription,
  //map<T>(fn: (_: E) => T): Event<T>,
  prepend<Before>(fn: (_: Before) => Params): Event<Before>,
  subscribe(subscriber: Subscriber<Params>): Subscription,
- to(store: Store<Params>, _: void): void,
- to<T>(store: Store<T>, reducer: (state: T, payload: Params) => T): void,
+ //prettier-ignore
+ +to: (
+  & (<T>(
+   store: Store<T>,
+   reducer: (state: T, payload: Params) => T
+  ) => Subscription)
+  & ((store: Store<Params>, _: void) => Subscription)
+ ),
  epic<T>(fn: (_: Stream<Params>) => Stream<T>): Event<T>,
  getType(): string,
  shortName: string,
@@ -77,20 +91,30 @@ export type Store<State> = {
  reset(event: Event<any> | Effect<any, any, any>): Store<State>,
  dispatch(action: any): any,
  getState(): State,
+ //prettier-ignore
+ +setState: (
+  & (<T>(newState: T, handler: (state: State, newState: T) => State) => void)
+  & (<T>(newState: State, _: void) => void)
+ ),
  withProps<Props, R>(
   fn: (state: State, props: Props) => R,
  ): (props: Props) => R,
- map<T>(fn: (_: State) => T): Store<T>,
+ map<T>(fn: (_: State, lastState?: T) => T): Store<T>,
  on<E>(
   event: Event<E> | Effect<E, any, any>,
-  handler: (state: State, payload: E) => State,
+  handler: (state: State, payload: E) => State | void,
  ): Store<State>,
+ to<T>(store: Store<T>, reducer: (state: T, payload: State) => T): Subscription,
  subscribe(listner: any): Subscription,
- watch<E>(watcher: (state: State, payload: E, type: string) => any): void,
+ thru<U>(fn: (store: Store<State>) => U): U,
+ watch<E>(
+  watcher: (state: State, payload: E, type: string) => any,
+ ): Subscription,
  graphite: {
   next: MultiStep,
   seq: SeqStep,
  },
+ displayName?: string,
 }
 
 export type Domain = {