+++ src/react/index.h.js
@@ -1,7 +1,7 @@
 //@flow
 
 import type {ComponentType, Node} from 'react'
-import {Store} from 'effector'
+import type {Store} from 'effector'
 
 export type StoreConsumer<State> = ComponentType<{
  children: (state: State) => Node,